MAJOR Update and Release - Some additions, function changes and code clean up
COMPLETELY removed the various testing lines in the while loop. While
loop now reflects how the sentry should operate. (Sentry will scan,
calculate the object distances from the sensors, and will "fire" at X
based on the distance).

In the while loop, the sentry response has been updated. It will respond
to objects within 16cm, rather than 30cm. It will now also disable the
motor through the A4988 when the sentry is "firing", then re-enable once
it's finished.

Added a new audio data file that will produce the sound of an empty clip
(for when the sentry is "out of ammo")

Updated the _FLASH.ld file. Changed the origins of all the audio files
to make room for the new audio file to be stored in memory.

Updated playAudio to be able to play the new audio data. Declared a new pointer
and wave size variables for the audio.

Disabled some debugging functions, can be re-enabled for future use when
needed.

Cleaned up code (got rid of a few remaining unused lines). Added some
comments to other areas for some additional clarity.